# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ODI API
  description: |
    REST API for the ODI (Open Document Indexer) service.
    Provides document upload, OCR, indexing, search, and retrieval.
  version: 0.1.0
  contact:
    name: ODI Project
    url: https://github.com/denysvitali/odi
servers:
  - url: http://127.0.0.1:8085
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /api/v1/search:
    post:
      summary: Search indexed documents
      description: |
        Executes a search query against the OpenSearch index.
        Supports pagination via `scrollId`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                description: Raw OpenSearch response; shape depends on the query.
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Search execution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/documents:
    get:
      summary: List documents
      description: Returns a list of indexed documents (OpenSearch-backed).
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                description: Raw OpenSearch response.
        "500":
          description: Query failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/documents/{id}:
    get:
      summary: Get a single document
      description: Retrieves an indexed document by its composite ID (`scanID_sequenceID`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f-]+_\d+$'
          example: "550e8400-e29b-41d4-a716-446655440000_1"
      responses:
        "200":
          description: Document found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          description: Invalid document ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/files/{scanID}/{sequenceId}:
    get:
      summary: Retrieve raw file
      description: Returns the original image or PDF page bytes.
      parameters:
        - name: scanID
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]+$'
        - name: sequenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/thumbnails/{id}:
    get:
      summary: Get thumbnail
      description: |
        Returns a thumbnail image. If the thumbnail does not exist,
        it is generated on demand from the original file.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f-]+_\d+$'
      responses:
        "200":
          description: Thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Original page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/thumbnails/process:
    post:
      summary: Trigger thumbnail backfill
      description: |
        Scans the index for pages missing thumbnails and generates them
        in the background.
      responses:
        "202":
          description: Backfill started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Backfill failed to start
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/reindex:
    get:
      summary: Get reindex status
      description: Returns the current or most recent in-memory reindex job status.
      responses:
        "200":
          description: Reindex status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReindexStatus"
    post:
      summary: Start reindex
      description: Starts a background job that reprocesses stored pages and rebuilds OpenSearch entries.
      responses:
        "202":
          description: Reindex started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReindexStatus"
        "409":
          description: Reindex already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Reindex unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/upload:
    post:
      summary: Upload documents
      description: |
        Accepts a multipart/form-data upload of images or PDFs.
        Each file is stored, OCR'd, and indexed.
        Maximum total request size is 200 MB. Optional sequenceIDs preserves source page identifiers for remote ingestion; omitted identifiers use sequenceOffset and one-based sequential numbering.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more image or PDF files
                sequenceIDs:
                  type: string
                  description: JSON array of distinct positive page identifiers, one per file (for example "[1,3]").
                sequenceOffset:
                  type: integer
                  minimum: 0
                  description: Legacy offset used only when sequenceIDs is omitted.
                scanID:
                  type: string
                  format: uuid
                  description: Optional source scan identifier; generated when omitted.
      responses:
        "200":
          description: Upload processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Bad request (no files, malformed form)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate document detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/search/facets:
    post:
      summary: Aggregate search filter facets
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  aggregations:
                    type: object
                  hits:
                    type: object
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
  /api/v1/reminders:
    get:
      summary: List upcoming document deadlines
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemindersResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: days
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 3650
          default: 90
  /api/v1/insights:
    get:
      summary: Aggregate document amounts
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: year
        in: query
        schema:
          type: integer
      - name: date_from
        in: query
        schema:
          type: string
      - name: date_to
        in: query
        schema:
          type: string
  /api/v1/insights.csv:
    get:
      summary: Export amount aggregates as CSV
      responses:
        '200':
          description: CSV amount aggregates
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: year
        in: query
        schema:
          type: integer
      - name: date_from
        in: query
        schema:
          type: string
      - name: date_to
        in: query
        schema:
          type: string
  /api/v1/documents/{id}/similar:
    get:
      summary: Find related documents
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/SimilarDocument'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document or share not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/documents/{id}/summary:
    post:
      summary: Generate or retrieve a cached local LLM summary
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                  keyFacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyFact'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Optional feature is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document or share not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/chat:
    post:
      summary: Ask the local LLM about matching documents
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  citations:
                    type: array
                    items:
                      type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Optional feature is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                companies:
                  type: array
                  items:
                    type: string
                dateFrom:
                  type: string
                dateTo:
                  type: string
                hasBarcode:
                  type: boolean
                title:
                  type: string
              required:
              - question
  /api/v1/shares:
    get:
      summary: List active document share links
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  shares:
                    type: array
                    items:
                      $ref: '#/components/schemas/Share'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Optional feature is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create an expiring document share link
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  urlSuffix:
                    type: string
                  expiresAt:
                    type: integer
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Optional feature is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scanID:
                  type: string
                sequenceID:
                  type: integer
                  minimum: 1
                expiresInHours:
                  type: integer
                  default: 24
                maxViews:
                  type: integer
                  minimum: 0
                passphrase:
                  type: string
                  writeOnly: true
              required:
              - scanID
              - sequenceID
  /api/v1/shares/{token}:
    delete:
      summary: Revoke a document share link
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Dependency or processing failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document or share not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
  /share/{token}:
    get:
      summary: Retrieve a document using a signed share token
      security: []
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      - name: X-Share-Passphrase
        in: header
        schema:
          type: string
          writeOnly: true
        description: Required when the share has a passphrase. Query-string passphrases
          are not accepted.
      responses:
        '200':
          description: Shared document
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '403':
          description: Passphrase required or incorrect
        '404':
          description: Share unavailable, expired, revoked, or exhausted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Set `API_TOKEN` on the server and send the value as a
        `Bearer` token in the `Authorization` header.
        If `API_TOKEN` is not configured, authentication is skipped.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "upload failed"
    SearchRequest:
      type: object
      properties:
        searchTerm:
          type: string
          example: "invoice"
        size:
          type: integer
          description: Page size (max 1000)
          maximum: 1000
          example: 20
        scrollId:
          type: string
          description: Pagination cursor from a previous search response
...
    DocumentResponse:
      type: object
      properties:
        _index:
          type: string
        _id:
          type: string
        _version:
          type: integer
        _seq_no:
          type: integer
        _primary_term:
          type: integer
        found:
          type: boolean
        _source:
          $ref: "#/components/schemas/Document"
    UploadResponse:
      type: object
      properties:
        scanID:
          type: string
          format: uuid
        processed:
          type: integer
        duplicates:
          type: integer
        failed:
          type: integer
        pages:
          type: array
          items:
            $ref: "#/components/schemas/UploadPageResult"
...